Fix QMutex::tryLock timeout regression (integer overflow)
The timeout given in milliseconds should be converted to a timespec
struct.
To separate the seconds and nanoseconds the milliseconds are first
multiplied to represent the whole value in an int64 variable.
The calculation is done on integers, so we get an overflow if the
milliseconds are bigger then 2148.
If we cast the given value to an int64 we can avoid this problem.

Fix the used cast.

Task-number: QTBUG-24795

+++ src/corelib/thread/qmutex_linux.cpp
@@ -121,7 +121,7 @@ bool QBasicMutex::lockInternal(int timeout)
                 struct timespec ts, *pts = 0;
                 if (timeout >= 1) {
                     // recalculate the timeout
-                    qint64 xtimeout = timeout * 1000 * 1000;
+                    qint64 xtimeout = qint64(timeout) * 1000 * 1000;
                     xtimeout -= elapsedTimer.nsecsElapsed();
                     if (xtimeout <= 0) {
                         // timer expired after we returned
